# docs/swagger.yaml
basePath: /ih/mobile_central_2_0_0
definitions:
...
  dto.PostChangePasswordDto:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  dto.PostMachineInfo:
    properties:
      machineId:
        type: string
    required:
    - machineId
    type: object
...
  dto.PostUserSetNewPasswordDto:
    properties:
      email:
        type: string
      password:
        type: string
      verificationCode:
        type: string
    required:
    - email
    - password
    - verificationCode
    type: object
...
  dto.SearchKeywordDto:
    properties:
      keyword:
        type: string
    type: object
...
  dto.UserEmailVerificationDto:
    properties:
      email:
        type: string
      verificationCode:
        type: string
    required:
    - email
    - verificationCode
    type: object
  dto.UserMobileVerificationDto:
    properties:
      mobileNo:
        type: string
      tac:
        type: string
    required:
    - mobileNo
    - tac
    type: object
...
  model.ApplicationUser:
    properties:
      address:
        type: string
      address1:
        type: string
      address2:
        type: string
      address3:
        type: string
      cityState:
        type: string
      contactNumber:
        type: string
      country:
        type: string
      dateLoggedIn:
        type: string
      dob:
        type: string
      docNoSignUp:
        type: string
      email:
        type: string
      firstName:
        type: string
      firstTimeBiometric:
        type: boolean
      firstTimeLogin:
        type: boolean
      fullName:
        type: string
      fullNameSignUp:
        type: string
      inactive:
        type: string
      isGoldenPearl:
        type: string
      isKidsExplorer:
        type: string
      isLoggedIn:
        type: integer
      lastName:
        type: string
      machineId:
        type: string
      masterPrn:
        type: string
      middleName:
        type: string
      nationality:
        type: string
      passport:
        type: string
      playerId:
        type: string
      postalCode:
        type: string
      race:
        type: string
      registerDate:
        type: string
      registerDateExcel:
        type: string
      resident:
        type: string
      role:
        type: string
      sessionId:
        type: string
      sex:
        type: string
      signInType:
        type: integer
      title:
        type: string
      user_id:
        type: integer
      userBranches:
        items:
          $ref: '#/definitions/model.AssignBranch'
        type: array
      username:
        type: string
      verificationCode:
        type: string
    type: object
...
  model.AssignBranch:
    properties:
      adminId:
        type: integer
      assignBranchId:
        type: integer
      branch:
        $ref: '#/definitions/model.Branch'
      branchId:
        type: integer
      prn:
        type: string
      userId:
        type: integer
    type: object
  model.Branch:
    properties:
      branchId:
        type: integer
      branchName:
        type: string
      passcode:
        type: string
      url:
        type: string
    type: object
...
  model.OnesignalNotification:
    properties:
      accountNo:
        type: string
      dateCreate:
        type: string
      dateSent:
        type: string
      fullMessage:
        type: string
      guestPlayerId:
        type: string
      isSeen:
        type: boolean
      master_id:
        type: integer
      msgType:
        type: string
      notification_id:
        type: integer
      notificationTitle:
        type: string
      shortMessage:
        type: string
      user_id:
        type: integer
      visitType:
        type: string
    type: object
...
info:
  contact: {}
  description: Vesalius-m Backend API Docs (IHP-UAT Version).
  title: (IHP-UAT) Vesalius-m Backend API
  version: "1.0"
paths:
...
  /user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ApplicationUser'
      security:
      - BearerAuth: []
      tags:
      - User
...
  /user/active-user/{userId}:
    post:
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/add-machine-id:
    post:
      parameters:
      - description: AddMachineId
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PostMachineInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/all:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationUser'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
    post:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      - description: Search
        in: body
        name: keyword
        schema:
          $ref: '#/definitions/dto.SearchKeywordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationUser'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
  /user/all/active:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationUser'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
  /user/branches:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AssignBranch'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
  /user/change-password:
    post:
      parameters:
      - description: PostChangePasswordDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PostChangePasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/delete-account:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/disable-firsttime-bio:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/export/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationUser'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
  /user/export/search:
    post:
      parameters:
      - description: Search
        in: body
        name: keyword
        schema:
          $ref: '#/definitions/dto.SearchKeywordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ApplicationUser'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
  /user/inactive-user/{userId}:
    post:
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/notification/list:
    get:
      parameters:
      - description: _page
        in: query
        name: _page
        type: string
      - description: _limit
        in: query
        name: _limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OnesignalNotification'
            type: array
      security:
      - BearerAuth: []
      tags:
      - User
  /user/notification/unseen/count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      security:
      - BearerAuth: []
      tags:
      - User
  /user/set-new-password:
    post:
      parameters:
      - description: PostUserSetNewPasswordDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PostUserSetNewPasswordDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - User
  /user/update-playerid/{playerId}:
    post:
      parameters:
      - description: playerId
        in: path
        name: playerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      tags:
      - User
  /user/userId/{userId}:
    get:
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ApplicationUser'
      security:
      - BearerAuth: []
      tags:
      - User
  /user/verify:
    post:
      parameters:
      - description: UserEmailVerificationDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserEmailVerificationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - User
  /user/verify-email:
    post:
      parameters:
      - description: UserEmailVerificationDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserEmailVerificationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - User
  /user/verify-smstac:
    post:
      parameters:
      - description: UserMobileVerificationDto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserMobileVerificationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"